Views system and Votes system have been fully implemented, added Search page with it's components, (i don't remember all the changes that have been made since the last commit)

+++ hbomax-backend/src/main/java/com/hbomax/models/Vote.java
@@ -14,7 +14,7 @@ public class Vote {
     @Column(name = "vote_id")
     private Integer voteId;
     @Column(name = "rating")
-    private Integer rating;
+    private float rating;
     @ManyToOne
     @JoinColumn(name = "user_vote", referencedColumnName = "user_id")
     private ApplicationUser userVote;
@@ -47,11 +47,12 @@ public class Vote {
         this.voteId = voteId;
     }
 
-    public Integer getRating() {
+
+    public float getRating() {
         return rating;
     }
 
-    public void setRating(Integer rating) {
+    public void setRating(float rating) {
         this.rating = rating;
     }
